Update router-Router documentation
After trying to work out for 30 minutes why my parameters defined in middleware were not being carried over, I eventually found a thread on GitHub that indicated the parameter only existed since `4.5.0` (strongloop/express#2257). As I was using `4.2.0`, this parameter didn't work for me and it would quietly fail.

This fix changes the documentation for the `mergeParams` parameter to clearly indicate that it has only existed since `4.5.0`.

Submitting as a GitHub edit rather than a regular pull request, because of local build issues (#297).