Update req-secure header and excerpt
Based on behavior of current site, it looks like req-secure ought to have the header reinstated. Also, es-lint fails with the JS block-style example for a few reasons, most importantly the "unused expression". For this example, I opted to just have a single line instead of the block-style code, which I am not sure could be made acceptable to es-lint without a less attractive option like introducing a variable (e.g. assign the result of the expression to isSecure...)

revert to block code excerpt with eslint exception

move eslint exception comment