Update error-handling.md
Inside the Catching Errors section, the sentence starting from line 64 to 65 it says 'If the callback in a sequence provides no data, only errors, you can simplify
this code as follows:'. This is little confusing for me. would it make more sense if the sentence is modified like this: 'If the callback in a sequence provides no data, only errors, you can simplify write code as follows:'. 

I might be wrong but please make me clear incase if so.