Fixed disappearing links
The links were diaappearing due to the toggle method called on the #navmenu ul in jqeury, which let to an inline styling of display:none . Now when we click on hamburger menu in smaller width two times it first adds display block and then changes to none on the second click. Now as it is an inline css we can override it using !important. 

Files changed:
1. css/style.css

Changes description:
Added display: block !important on screen width being more than equal to 900px