Added number of hops to function example for "behind-proxies" page
`proxy-addr` not only passes the ip but also the index of the hop to the callback. This could be noted in the documentation, as it's useful for private servers behind a public load balancer which, in theory, is accessible to the public internet but should be proxied by a CDN.
In this example the first hop can always be trusted (LoadBalancer -> API) but the second hop would depend on the IP of the incoming request (either CDN -> LB or internet -> LB).