Display author avatars in the blog (#1915)
* feat: allow blogs to have multiple authors

This requires us to update the `author` front-matter, requiring it to be
an array of objects instead of a single string. This paves the way for
us to include further author metadata.

For now we only add "name" which expects a string.

* feat: allow linking to author GitHub profiles

This adds the ability for authors to add their GitHub username to the
blog posts they author. If they provide this then their name will be
linked to their GitHub profile page.

* feat: add GitHub profile images to author lines

This adds a small profile image to author lines that have a GitHub.

* feat: add author URLs to the feeds

* fix: space the authors a little more

The authors looked a little bunched up, a left margin on each of them
reads a little better.

Co-Authored-By: Wes Todd <wesleytodd@users.noreply.github.com>

* fix: improve readability of the feed entry Liquid

Co-Authored-By: Sebastian Beltran <bjohansebas@users.noreply.github.com>

* feat: add author image for the technical committee

Co-Authored-By: Sebastian Beltran <bjohansebas@users.noreply.github.com>

* fix: set a width and height on author images

This ensures we still display a circle if, for some reason, the image
that we get isn't a perfect square.

Co-Authored-By: Shubham Oulkar <ShubhamOulkar@users.noreply.github.com>

---------

Co-authored-by: Wes Todd <wesleytodd@users.noreply.github.com>
Co-authored-by: Sebastian Beltran <bjohansebas@users.noreply.github.com>
Co-authored-by: Shubham Oulkar <ShubhamOulkar@users.noreply.github.com>